Move NewGroupChatModal.jsx SelectContactsItem.jsx SelectContactsList.jsx Modals/NewGroupChatModal -> Modals
