+++ frontend/src/App.jsx
@@ -25,7 +25,7 @@ import Settings from "./components/Settings/Settings";
 
 import NewChatDropDownBox from "./components/Modals/NewChatDropDownBox";
 import NewPrivateChatModal from "./components/Modals/NewPrivateChatModal/NewPrivateChatModal";
-import NewGroupChatModal from "./components/Modals/NewGroupChatModal/NewGroupChatModal";
+import NewGroupChatModal from "./components/Modals/NewGroupChatModal";
 import AddNewContactsModal from "./components/Modals/AddNewContactsModal";
 
 const App = () => {
