+++ frontend/src/components/Modals/AddNewContactsModal.jsx
@@ -14,7 +14,7 @@ import useField from "../../hooks/useField";
 
 import Loading from "../Loading";
 import SearchBar from "../SearchBar";
-import SelectContactsList from "./NewGroupChatModal/SelectContactsList";
+import SelectContactsList from "./SelectContactsList";
 
 import useNotifyMessage from "../../hooks/useNotifyMessage";
 import Notify from "../Notify";
