+++ frontend/src/components/Modals/EditGroupChatModal/UpdateMembersModal.jsx
@@ -10,7 +10,7 @@ import useField from "../../../hooks/useField";
 
 import Loading from "../../Loading";
 import SearchBar from "../../SearchBar";
-import SelectContactsList from "../NewGroupChatModal/SelectContactsList";
+import SelectContactsList from "../SelectContactsList";
 
 const UpdateMembersModal = ({
   chat,
